refactor: move WeekRangeFilter to view.component and fix language-observer leak
